Adds Configuration object and wizard object to create config file.
The steps needed to create a config file have been extracted to
CarthageCache::ConfiguratorWizard. Also the configuration is
now represented with a CarthageCache::Configuration object instead
of a hash.

If the CARTHAGE_CACHE_DEFAULT_BUCKET_NAME env variable is defined
when running 'carthage_cache config', when the user is prompted
to enter the bucket name it will the env variable value as
the default value in case he/she wants to skip that step.